reorganize for reading in
