+++ environment.yml
@@ -6,6 +6,7 @@ dependencies:
    - python=3.8
    - pip
    - ipython
+   - ipykernel
    # Other packages
    - pandas
    - numpy
